상품 데이터 삭제 및 디렉토리 삭제

+++ admin/products/pro_delete.php
@@ -12,34 +12,20 @@ $page = set_var($_GET['page']);
 $query  = "SELECT * FROM products WHERE num='$p_num'";
 $result = mysqli_query($connect, $query);
 $row    = mysqli_fetch_array($result);
-mysqli_free_result($result);
 
-//작은 이미지가 있으면 삭제
-if ($row['s_image1'] == 'Y') {
-    unlink($row['s_image1_name']);
-}
-if ($row['b_image1'] == 'Y') {
-    unlink($row['b_image1_name']);
-}
-if ($row['b_image2'] == 'Y') {
-    unlink($row['b_image2_name']);
-}
-if ($row['b_image3'] == 'Y') {
-    unlink($row['b_image3_name']);
-}
-if ($row['b_image4'] == 'Y') {
-    unlink($row['b_image4_name']);
-}
-if ($row['b_image5'] == 'Y') {
-    unlink($row['b_image5_name']);
-}
-if ($row['d_image'] == 'Y') {
-    unlink($row['d_image_name']);
-}
+// ex) $imgPath = "../../upload/p_image/B068-02/b/4848_3.jpg"
+// $dir = "../../upload/p_image/B068-02"
+$imgPath = $row['b_image1_name'];
+$dir     = reverse_strrchr($imgPath, '/');
 
-// 해당데이타 삭제
-$query = "DELETE FROM products WHERE num='$p_num' ";
-mysqli_query($connect, $query);
+// 서브디렉토리까지 모두 삭제
+if (recurseRmdir($dir)) {
+
+    // 해당 상품데이타 삭제
+    $query = "DELETE FROM products WHERE num='$p_num' ";
+    mysqli_query($connect, $query);
+    msg("상품을 삭제했습니다.");
+}
 
 // 리스트로 이동
 // echo ("<meta http-equiv='refresh' content='0; URL=top_pro_list.php?lcode=$lcode&amp;mcode=$mcode&amp;page=$page'>");
